BREAKING CHANGE: Move matrix functions into namespace
- `vg.pad_with_ones()` -> `vg.matrix.pad_with_ones()`
- `vg.unpad()` -> `vg.matrix.unpad()`
- `vg.apply_homogeneous()` -> `vg.matrix.transform()`
