+++ vg/core.py
@@ -1,5 +1,6 @@
 import math
 import numpy as np
+from ._helpers import pluralize, raise_dimension_error
 
 __all__ = [
     "normalize",
@@ -14,34 +15,12 @@ __all__ = [
     "rotate",
     "almost_zero",
     "almost_collinear",
-    "pad_with_ones",
-    "unpad",
-    "apply_homogeneous",
     "principal_components",
     "major_axis",
     "apex",
     "farthest",
     "basis",
 ]
-
-
-def pluralize(noun, count):
-    return noun if count == 1 else "{}s".format(noun)
-
-
-def raise_dimension_error(*input_values):
-    messages = [
-        "{} {}".format(input_value.ndim, pluralize("dimension", input_value.ndim))
-        for input_value in input_values
-    ]
-    if len(messages) == 1:
-        message = messages[0]
-    elif len(messages) == 2:
-        message = "{} and {}".format(*messages)
-    else:
-        message = "those inputs"
-    raise ValueError("Not sure what to do with {}".format(message))
-
 
 def normalize(vector):
     """
@@ -341,65 +320,6 @@ def almost_collinear(v1, v2, atol=1e-08):
     cross = np.cross(v1, v2)
     norm = np.linalg.norm(cross)
     return np.isclose(norm, 0.0, rtol=0, atol=atol)
-
-
-def pad_with_ones(matrix):
-    """
-    Add a column of ones. Transform from:
-        array([[1., 2., 3.],
-               [2., 3., 4.],
-               [5., 6., 7.]])
-    to:
-        array([[1., 2., 3., 1.],
-               [2., 3., 4., 1.],
-               [5., 6., 7., 1.]])
-
-    """
-    if matrix.ndim != 2 or matrix.shape[1] != 3:
-        raise ValueError("Invalid shape %s: pad expects nx3" % (matrix.shape,))
-    return np.pad(matrix, ((0, 0), (0, 1)), mode="constant", constant_values=1)
-
-
-def unpad(matrix):
-    """
-    Strip off a column (e.g. of ones). Transform from:
-        array([[1., 2., 3., 1.],
-               [2., 3., 4., 1.],
-               [5., 6., 7., 1.]])
-    to:
-        array([[1., 2., 3.],
-               [2., 3., 4.],
-               [5., 6., 7.]])
-
-    """
-    if matrix.ndim != 2 or matrix.shape[1] != 4:
-        raise ValueError("Invalid shape %s: unpad expects nx4" % (matrix.shape,))
-    if not all(matrix[:, 3] == 1.0):
-        raise ValueError("Expected a column of ones")
-    return np.delete(matrix, 3, axis=1)
-
-
-def apply_homogeneous(vertices, transform):
-    """
-    Apply the given transformation matrix to the vertices using homogenous
-    coordinates.
-    """
-    if transform.shape != (4, 4):
-        raise ValueError("Transformation matrix should be 4x4")
-
-    if vertices.ndim == 1:
-        matrix = vertices[np.newaxis]
-    elif vertices.ndim == 2:
-        matrix = vertices
-    else:
-        raise_dimension_error(vertices)
-
-    if matrix.shape[1] != 3:
-        raise ValueError("Vertices should be 3x1 or Nx3")
-
-    result = unpad(np.dot(transform, pad_with_ones(matrix).T).T)
-    return result[0] if vertices.ndim == 1 else result
-
 
 def principal_components(coords):
     """
