+++ vg/test_matrix_transform.py
@@ -1,6 +1,6 @@
 import numpy as np
 import pytest
-from . import core as vg
+from .matrix import transform as apply_transform
 
 scale_factor = np.array([3.0, 0.5, 2.0])
 transform = np.array(
@@ -17,7 +17,7 @@ def test_apply_homogeneous():
     point = np.array([5.0, 0.0, 1.0])
     expected_point = np.array([15.0, 0.0, 2.0])
     np.testing.assert_array_equal(
-        vg.apply_homogeneous(point, transform), expected_point
+        apply_transform(point, transform), expected_point
     )
 
 
@@ -25,14 +25,14 @@ def test_apply_homogeneous_stacked():
     points = np.array([[1.0, 2.0, 3.0], [5.0, 0.0, 1.0]])
     expected_points = np.array([[3.0, 1.0, 6.0], [15.0, 0.0, 2.0]])
     np.testing.assert_array_equal(
-        vg.apply_homogeneous(points, transform), expected_points
+        apply_transform(points, transform), expected_points
     )
 
 
 def test_apply_homogeneous_error():
     with pytest.raises(ValueError, match="Transformation matrix should be 4x4"):
-        vg.apply_homogeneous(np.array([1.0, 2.0, 3.0]), np.array([1.0]))
+        apply_transform(np.array([1.0, 2.0, 3.0]), np.array([1.0]))
     with pytest.raises(ValueError, match="Vertices should be 3x1 or Nx3"):
-        vg.apply_homogeneous(np.array([1.0, 2.0]), transform)
+        apply_transform(np.array([1.0, 2.0]), transform)
     with pytest.raises(ValueError, match="Not sure what to do with 3 dimensions"):
-        vg.apply_homogeneous(np.array([[[1.0, 2.0, 3.0]]]), transform)
+        apply_transform(np.array([[[1.0, 2.0, 3.0]]]), transform)
